+++ CHANGELOG.md
@@ -12,7 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - [#12](https://github.com/LDeakin/zarrs_tools/pull/12) Bump netcdf to 0.10.2 by [@magnusuMET]
  - **Breaking**: Bump MSRV to 1.80
  - Bump `sysinfo` to 0.31
- - Bump `zarrs` to 0.17.0-beta.2
+ - Bump `zarrs` to 0.17.0-beta.3
+ - `ncvar2zarr`:
+   - Switch to output concurrency
+   - **Breaking**: Rename `concurrent-blocks` argument to `concurrent-chunks`
+   - **Breaking**: Removed `validate` argument
+   - **Breaking**: Removed `concat-dim` argument. Concatenation is now only supported along the first (slowest varying) dimension
+   - This tool is intended to be replaced by chunk manifests when the specification and `virtualizarr` matures
 
 ## [0.5.5] - 2024-07-31